Show the correct matching rule type
Now that the values of the rules are wrapped in an array an need to
be read from the name key. The correct value type was no longer shown
in the fourth column of the ARP attribute table. This was fixed by
loading the value using the `getAttributeFilterValue` helper method
specifically designed for this task (with backwards compatibility in
mind).